requirements.txt
coverage==5.5
Flask==2.0.1
flask-apispec==0.11.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
pytest==6.2.4
python-dotenv==0.19.0